+++ src/db/schema/index.ts
@@ -4,6 +4,7 @@ export * from "./parent";
 export * from "./auth";
 export * from "./curriculum";
 export * from "./calendar";
+export * from "./tutor";
 export * from "./textbook";
 export * from "./school-profile";
 export * from "./vocab";